Space Integration Tests Complete

+++ 09_Capstone/src/main/java/com/techelevator/space/Space.java
@@ -1,5 +1,7 @@
 package com.techelevator.space;
 
+import java.math.BigDecimal;
+
 public class Space {
 	
 	private long spaceId;
@@ -8,7 +10,7 @@ public class Space {
 	private boolean isAccessible;
 	private int openFrom;
 	private int openTo;
-	private double dailyRate;
+	private BigDecimal dailyRate;
 	private int maxOccupancy;
 	
 	public long getSpaceId() {
@@ -47,10 +49,10 @@ public class Space {
 	public void setOpenTo(int openTo) {
 		this.openTo = openTo;
 	}
-	public double getDailyRate() {
+	public BigDecimal getDailyRate() {
 		return dailyRate;
 	}
-	public void setDailyRate(double dailyRate) {
+	public void setDailyRate(BigDecimal dailyRate) {
 		this.dailyRate = dailyRate;
 	}
 	public int getMaxOccupancy() {
